fix MetaMaster auth
