+++ dora/core/client/fs/src/main/java/alluxio/client/meta/MetaMasterConfigClient.java
@@ -11,17 +11,12 @@
 
 package alluxio.client.meta;
 
-import alluxio.AlluxioURI;
 import alluxio.Client;
-import alluxio.conf.PropertyKey;
 import alluxio.grpc.GetConfigurationPOptions;
 import alluxio.wire.ConfigHash;
 import alluxio.wire.Configuration;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
 
 /**
  * Interface for a meta master config client.
@@ -37,51 +32,4 @@ public interface MetaMasterConfigClient extends Client {
    * @return hashes of cluster and path level configurations
    */
   ConfigHash getConfigHash() throws IOException;
-
-  /**
-   * Sets a property for a path.
-   *
-   * @param path the path
-   * @param key the property key
-   * @param value the property value
-   */
-  default void setPathConfiguration(AlluxioURI path, PropertyKey key, String value)
-      throws IOException {
-    Map<PropertyKey, String> properties = new HashMap<>();
-    properties.put(key, value);
-    setPathConfiguration(path, properties);
-  }
-
-  /**
-   * Sets properties for a path.
-   *
-   * @param path the path
-   * @param properties the properties
-   */
-  void setPathConfiguration(AlluxioURI path, Map<PropertyKey, String> properties)
-      throws IOException;
-
-  /**
-   * Removes properties for a path.
-   *
-   * @param path the path
-   * @param keys the property keys
-   */
-  void removePathConfiguration(AlluxioURI path, Set<PropertyKey> keys) throws IOException;
-
-  /**
-   * Removes all properties for a path.
-   *
-   * @param path the path
-   */
-  void removePathConfiguration(AlluxioURI path) throws IOException;
-
-  /**
-   * Updates properties.
-   *
-   * @param propertiesMap the properties map to be updated
-   * @return the update properties status map
-   */
-  Map<PropertyKey, Boolean> updateConfiguration(
-      Map<PropertyKey, String> propertiesMap) throws IOException;
 }
