+++ dora/core/client/fs/src/main/java/alluxio/client/meta/RetryHandlingMetaMasterClient.java
@@ -12,7 +12,9 @@
 package alluxio.client.meta;
 
 import alluxio.AbstractMasterClient;
+import alluxio.AlluxioURI;
 import alluxio.Constants;
+import alluxio.conf.PropertyKey;
 import alluxio.grpc.BackupPRequest;
 import alluxio.grpc.BackupStatusPRequest;
 import alluxio.grpc.CheckpointPOptions;
@@ -23,7 +25,10 @@ import alluxio.grpc.MasterInfo;
 import alluxio.grpc.MasterInfoField;
 import alluxio.grpc.MetaMasterClientServiceGrpc;
 import alluxio.grpc.ProxyStatus;
+import alluxio.grpc.RemovePathConfigurationPRequest;
 import alluxio.grpc.ServiceType;
+import alluxio.grpc.SetPathConfigurationPRequest;
+import alluxio.grpc.UpdateConfigurationPRequest;
 import alluxio.master.MasterClientContext;
 import alluxio.wire.BackupStatus;
 import alluxio.wire.ConfigCheckReport;
@@ -32,7 +37,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
 import javax.annotation.concurrent.ThreadSafe;
@@ -115,5 +123,50 @@ public class RetryHandlingMetaMasterClient extends AbstractMasterClient
     return retryRPC(() -> mClient.listProxyStatus(
         ListProxyStatusPRequest.getDefaultInstance()).getProxyStatusesList(),
         RPC_LOG, "ListProxyStatus", "");
+  }
+
+  @Override
+  public void setPathConfiguration(AlluxioURI path, Map<PropertyKey, String> properties)
+      throws IOException {
+    Map<String, String> props = new HashMap<>();
+    properties.forEach((key, value) -> props.put(key.getName(), value));
+    retryRPC(() -> mClient.setPathConfiguration(SetPathConfigurationPRequest.newBuilder()
+            .setPath(path.getPath()).putAllProperties(props).build()),
+        RPC_LOG, "setPathConfiguration", "path=%s,properties=%s", path, properties);
+  }
+
+  @Override
+  public void removePathConfiguration(AlluxioURI path, Set<PropertyKey> keys) throws IOException {
+    Set<String> keySet = new HashSet<>();
+    for (PropertyKey key : keys) {
+      keySet.add(key.getName());
+    }
+    retryRPC(() -> mClient.removePathConfiguration(RemovePathConfigurationPRequest.newBuilder()
+            .setPath(path.getPath()).addAllKeys(keySet).build()),
+        RPC_LOG, "removePathConfiguration", "path=%s,keys=%s", path, keys);
+  }
+
+  @Override
+  public void removePathConfiguration(AlluxioURI path) throws IOException {
+    retryRPC(() -> mClient.removePathConfiguration(RemovePathConfigurationPRequest.newBuilder()
+        .setPath(path.getPath()).build()), RPC_LOG, "removePathConfiguration", "path=%s", path);
+  }
+
+  @Override
+  public Map<PropertyKey, Boolean> updateConfiguration(
+      Map<PropertyKey, String> propertiesMap) throws IOException {
+    Map<PropertyKey, Boolean> resultMap = new HashMap<>();
+    Map<String, String> inputMap = new HashMap<>();
+    propertiesMap.forEach((k, v) -> inputMap.put(k.getName(), v));
+    retryRPC(
+        () -> mClient.updateConfiguration(
+            UpdateConfigurationPRequest.newBuilder()
+                .putAllProperties(inputMap)
+                .build()),
+        RPC_LOG, "updateConfiguration", "propertiesMap=%s", propertiesMap)
+        .getStatusMap().forEach((k, v) -> {
+          resultMap.put(PropertyKey.Builder.stringBuilder(k).buildUnregistered(), v);
+        });
+    return resultMap;
   }
 }
