+++ dora/core/server/master/src/main/java/alluxio/master/meta/MetaMasterConfigurationServiceHandler.java
@@ -12,28 +12,16 @@
 package alluxio.master.meta;
 
 import alluxio.RpcUtils;
-import alluxio.conf.PropertyKey;
 import alluxio.grpc.GetConfigHashPOptions;
 import alluxio.grpc.GetConfigHashPResponse;
 import alluxio.grpc.GetConfigurationPOptions;
 import alluxio.grpc.GetConfigurationPResponse;
 import alluxio.grpc.MetaMasterConfigurationServiceGrpc;
-import alluxio.grpc.RemovePathConfigurationPRequest;
-import alluxio.grpc.RemovePathConfigurationPResponse;
-import alluxio.grpc.SetPathConfigurationPRequest;
-import alluxio.grpc.SetPathConfigurationPResponse;
-import alluxio.grpc.UpdateConfigurationPRequest;
-import alluxio.grpc.UpdateConfigurationPResponse;
 import alluxio.wire.ConfigHash;
 
 import io.grpc.stub.StreamObserver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
 
 /**
  * This class is a gRPC handler for meta master RPCs.
@@ -96,48 +84,5 @@ public final class MetaMasterConfigurationServiceHandler
     RpcUtils.call(LOG, () ->
         mMetaMaster.getConfigHash().toProto(), "getConfigHash", "request=%s", responseObserver,
         request);
-  }
-
-  @Override
-  public void setPathConfiguration(SetPathConfigurationPRequest request,
-      StreamObserver<SetPathConfigurationPResponse> responseObserver) {
-    String path = request.getPath();
-    Map<String, String> properties = request.getPropertiesMap();
-
-    RpcUtils.call(LOG, () -> {
-      Map<PropertyKey, String> props = new HashMap<>();
-      properties.forEach((key, value) -> props.put(PropertyKey.fromString(key), value));
-      mMetaMaster.setPathConfiguration(path, props);
-      return SetPathConfigurationPResponse.getDefaultInstance();
-    }, "setPathConfiguration", "request=%s", responseObserver, request);
-  }
-
-  @Override
-  public void removePathConfiguration(RemovePathConfigurationPRequest request,
-      StreamObserver<RemovePathConfigurationPResponse> responseObserver) {
-    String path = request.getPath();
-    List<String> keys = request.getKeysList();
-
-    RpcUtils.call(LOG, () -> {
-      if (keys.isEmpty()) {
-        mMetaMaster.removePathConfiguration(path);
-      } else {
-        mMetaMaster.removePathConfiguration(path, new HashSet<>(keys));
-      }
-      return RemovePathConfigurationPResponse.getDefaultInstance();
-    }, "removePathConfiguration", "request=%s", responseObserver, request);
-  }
-
-  @Override
-  public void updateConfiguration(
-      UpdateConfigurationPRequest request,
-      StreamObserver<UpdateConfigurationPResponse> responseObserver) {
-    RpcUtils.call(LOG, () -> {
-      Map<String, Boolean> result =
-          mMetaMaster.updateConfiguration(request.getPropertiesMap());
-      return UpdateConfigurationPResponse.newBuilder()
-          .putAllStatus(result)
-          .build();
-    }, "updateConfiguration", "request=%s", responseObserver, request);
   }
 }
