+++ dora/tests/src/test/java/alluxio/client/cli/fsadmin/pathconf/ShowCommandIntegrationTest.java
@@ -17,8 +17,8 @@ import alluxio.client.ReadType;
 import alluxio.client.WriteType;
 import alluxio.client.cli.fs.AbstractShellIntegrationTest;
 import alluxio.client.file.FileSystemContext;
-import alluxio.client.meta.MetaMasterConfigClient;
-import alluxio.client.meta.RetryHandlingMetaMasterConfigClient;
+import alluxio.client.meta.MetaMasterClient;
+import alluxio.client.meta.RetryHandlingMetaMasterClient;
 import alluxio.conf.Configuration;
 import alluxio.conf.InstancedConfiguration;
 import alluxio.conf.PropertyKey;
@@ -52,7 +52,7 @@ public class ShowCommandIntegrationTest extends AbstractShellIntegrationTest {
    */
   private InstancedConfiguration setPathConfigurations() throws Exception {
     FileSystemContext metaCtx = FileSystemContext.create(Configuration.global());
-    MetaMasterConfigClient client = new RetryHandlingMetaMasterConfigClient(
+    MetaMasterClient client = new RetryHandlingMetaMasterClient(
         MasterClientContext.newBuilder(metaCtx.getClientContext()).build());
     client.setPathConfiguration(new AlluxioURI(DIR1), PROPERTY_KEY11, PROPERTY_VALUE11);
     client.setPathConfiguration(new AlluxioURI(DIR1), PROPERTY_KEY12, PROPERTY_VALUE12);
